ff_ffplay: merge: 5edab1d207d211962b8b0e9ffdebf774aa6ee688
cmdutils: remove sws_opts usage, simplify code

It has become unused as all code was switched to AVDictionary

+++ ijkmedia/ijkplayer/ff_ffplay_def.h
@@ -118,7 +118,7 @@
 #ifdef FFP_MERGE
 #define CURSOR_HIDE_DELAY 1000000
 
-static int64_t sws_flags = SWS_BICUBIC;
+static unsigned sws_flags = SWS_BICUBIC;
 #endif
 
 typedef struct MyAVPacketList {
@@ -501,7 +501,7 @@ typedef struct FFPlayer {
 #endif
     int autorotate;
 
-    int sws_flags;
+    unsigned sws_flags;
 
     /* current context */
 #ifdef FFP_MERGE
